requirements.txt
flask
flask-jwt-extended
flask-cors
mysql-connector-python
python-dotenv
apscheduler
werkzeug
